revised label qdt.

+++ dataset/s8d_2d.py
@@ -351,7 +351,7 @@ class S8DFinetune2DAP(Dataset):
         
         # Convert to tensors
         seq_img = torch.from_numpy(seq_img).float()  # Add channel dim
-        seq_mask = torch.from_numpy(seq_mask).long()
+        seq_mask = torch.from_numpy(seq_mask)
         seq_img = (seq_img - seq_img.min()) / (seq_img.max() - seq_img.min()+1e-4)
         import pdb;pdb.set_trace()
         seq_mask = F.one_hot(seq_mask, num_classes=self.num_classes)
